removing duplicate enricher
