+++ helm/robusta/values.yaml
@@ -155,7 +155,7 @@ builtinPlaybooks:
   - on_prometheus_alert:
         alert_name: PrometheusRuleFailures
   actions:
-  - prometheus_logs_enricher:
+  - logs_enricher:
       lines_with_regex: ".*Evaluating rule failed.*"
 
 - triggers:
