+++ playbooks/robusta_playbooks/alerts_integration.py
@@ -1,4 +1,5 @@
 import logging
+import re
 import time
 from collections import defaultdict
 from datetime import datetime
@@ -318,6 +319,7 @@ class LogEnricherParams(ActionParams):
     :var warn_on_missing_label: Send a warning if the alert doesn't have a pod label
     :var regex_replacer_patterns: regex patterns to replace text, for example for security reasons (Note: Replacements are executed in the given order)
     :var regex_replacement_style: one of SAME_LENGTH_ASTERISKS or NAMED (See RegexReplacementStyle)
+    :var lines_with_regex: only shows lines that match the regex
     """
 
     container_name: Optional[str]
@@ -325,6 +327,12 @@ class LogEnricherParams(ActionParams):
     regex_replacer_patterns: Optional[List[NamedRegexPattern]] = None
     regex_replacement_style: Optional[str] = None
     previous: bool = False
+    lines_with_regex: Optional[str] = None
+
+
+def filter_logs(logs: str, lines_with_regex: str) -> str:
+    regex = re.compile(lines_with_regex)
+    return "\n".join(re.findall(regex, logs))
 
 
 @action
@@ -368,7 +376,8 @@ def logs_enricher(event: PodEvent, params: LogEnricherParams):
             logging.info("log data is empty, retrying...")
             time.sleep(backoff_seconds)
             continue
-
+        if params.lines_with_regex:
+            log_data = filter_logs(log_data, params.lines_with_regex)
         event.add_enrichment(
             [FileBlock(f"{pod.metadata.name}/{container}.log", log_data.encode())],
         )
